Chorizo vdev ChorizoCore.Authorization behaviour

This module contains many implementations of authorized/3, which are used to determine if a given user has a given permission.

Types

permission()

permission() :: atom()

A named permission used throughout the system

t()

t() :: module()

The module implementing the authorization logic. Must implement the ChorizoCore.Repositories.Authorization behaviour. Defaults to ChorizoCore.Repositories.Authorization.

Functions

authorized?(permission, user, users_repo \\ Users)

authorized?(permission(), ChorizoCore.Entities.User.t(), ChorizoCore.Repositories.Users.t()) :: boolean()

Returns true if the user is granted the specified permission

Allowing the first user to be created

When there are no users in the users_repo, the anonymous user is granted :manage_users. This is to allow the first user to be created.

Callbacks

authorized?(permission, arg1, arg2)

authorized?(permission(), ChorizoCore.Entities.User.t(), ChorizoCore.Repositories.Users.t()) :: boolean()