Chorizo vdev ChorizoCore.Entities.User

Provides functions for working with user data

Types

t()

t() :: %ChorizoCore.Entities.User{__meta__: term(), admin: boolean(), anonymous: boolean(), id: String.t() | nil, inserted_at: term(), password: String.t() | nil, password_hash: String.t() | nil, updated_at: term(), username: String.t() | nil}

Contains the data related to an individual user of the system

Functions

anonymous!()

anonymous!() :: t()

Builds and returns an anonymous %ChorizoCore.Entities.User{}

changeset(params)

changeset(%{}) :: Ecto.Changeset.t()

Builds an Ecto.Changeset for the user data

changeset(user, params \\ %{})

changeset(ChorizoCore.Entities.User.t(), %{}) :: Ecto.Changeset.t()

Builds an Ecto.Changeset for the user data

Note that if a :password key is present in the params, its value will be hashed with ChorizoCore.Authentication.Hasher.hashpwsalt/1 and stored in the :password_hash key. The :password key will be removed.

is_anonymous(user)

is_anonymous(t()) :: boolean()

Determines if the User is anonymous