Chorizo vdev ChorizoCore.Repositories.API behaviour

Behavior module for the Repositories API

Types

error()

error() :: {:error, term()}

Returned by functions that encounter an error during operation, e.g. if an attempt is made to insert an invalid record.

found(type)

found(type) :: {:ok, type}

Returned by functions that successfully find a single result

found_multi(type)

found_multi(type) :: {:ok, [type, ...]}

Returned by functions that successfully find multiple results

multi_result(type)

multi_result(type) :: found_multi(type) | not_found([]) | error()

Return value for functions that are expected to return many results

not_found(type)

not_found(type) :: {:not_found, type}

Returned by functions that execute successfully but do not find any records

record()

record() :: struct()

The type of record stored by the repository

search_options()

search_options() :: keyword()

Search options used by the repository to filter results

single_result(type)

single_result(type) :: found(type) | not_found(nil) | error()

Return value for functions that are expected to return at most 1 result

Callbacks

count()

count() :: single_result(integer())

Returns the total number of items in the repository

count(search_options)

count(search_options()) :: single_result(integer())

Returns the number of items in the repository matching search_options

first() (optional)

first() :: single_result(record())

Finds the first item in the repository

first(search_options)

first(search_options()) :: single_result(record())

Finds the first item in the repository that matches search_options

insert(record)

insert(record()) :: single_result(record())

Inserts record into the repository